added gerar comprovante functionalities

+++ database/migrations/2021_11_24_030523_create_ctes_table.php
@@ -16,9 +16,9 @@ class CreateCtesTable extends Migration
         Schema::create('ctes', function (Blueprint $table) {
             $table->id();
             $table->string('numero_nf', 20)->unique();
-            $table->double('valor_nf', 8, 2);
+            $table->integer('valor_nf');
             $table->string('numero_cte', 20)->unique();
-            $table->double('valor_cte', 8, 2);
+            $table->integer('valor_cte');
             $table->date('data_chegada');
             $table->date('data_entrega')->nullable();
             $table->enum('tipo_pagamento', ['CIF', 'FOB']);
